% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycan_structure.R
\name{compute_structural_traits}
\alias{compute_structural_traits}
\title{Compute structural glycan traits from an igraph glycan tree}
\usage{
compute_structural_traits(tree)
}
\arguments{
\item{tree}{A parsed glycan tree from \code{\link{pGlyco3_to_tree}} or \code{\link{wurcs_to_tree}}.}
}
\value{
A named numeric vector of built-in structural traits and
user-defined motif counts
}
\description{
Evaluate build-in structural glycan traits including:
\itemize{
  \item \code{Antenna numbers}
  \item \code{Bisecting-type}
  \item \code{Complex-type}
  \item \code{High-mannose-type}
  \item \code{Hybrid-type}
  \item \code{Core-fucosed}
  \item \code{Antennary-fucosed}
}
}
\keyword{internal}
