% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toyexample.R
\docType{data}
\name{glycanDatabase}
\alias{glycanDatabase}
\title{Glycan annotation reference database}
\format{
A data frame with N rows and M columns.
}
\usage{
data(glycanDatabase)
}
\description{
A curated reference table mapping glycan IDs to their structures,
used internally by glycoTraitR and also available to users.
}
\examples{
data(glycanDatabase)
head(glycanDatabase)
}
\keyword{datasets}
