% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glycoTraitR-pacakge.R
\docType{package}
\name{glycoTraitR-package}
\alias{glycoTraitR}
\alias{glycoTraitR-package}
\title{glycoTraitR: Compute and analyze the glycan structrual traits from GPSM data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

GlycoTraitR is an R package for analyzing glycoproteomics data, particularly glycopeptide-spectrum matches (GPSMs). It supports results generated by the pGlyco3 and Glyco-Decipher search engines. The package parses glycan structures, computes monosaccharide compositions and structural traits, quantifies glycan heterogeneity at the glycosylation site and protein levels, supports user-defined structural motifs, and provides visualization utilities for interpreting glycan trait changes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/matsui-lab/glycoTraitR}
  \item Report bugs at \url{https://github.com/matsui-lab/glycoTraitR/issues}
}

}
\author{
\strong{Maintainer}: Bingyuan Zhang \email{zhang.bingyuan.w8@f.mail.nagoya-u.ac.jp} (\href{https://orcid.org/0000-0002-4892-323X}{ORCID})

Authors:
\itemize{
  \item Koichi Himori \email{himori.koichi.b5@f.mail.nagoya-u.ac.jp}
  \item Yusuke Matsui \email{matsui.yusuke.d4@f.mail.nagoya-u.ac.jp} [funder]
}

}
\keyword{internal}
