% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toyexample.R
\docType{data}
\name{meta_toyexample}
\alias{meta_toyexample}
\title{Toy metadata for glycoTraitR examples}
\format{
A data frame with 34 rows and 2 variables:
\describe{
  \item{Diagnosis}{Clinical diagnosis (factor)}
  \item{Sample number}{Sample ID (numeric)}
  \item{file}{File Name (character)}
}
}
\usage{
data(meta_toyexample)
}
\description{
A toy example metadata table used in vignettes, examples, and tests.
}
\examples{
data(meta_toyexample)
head(meta_toyexample)
}
\keyword{datasets}
